+++ .vscode/arduino.json
@@ -1,6 +1,6 @@
 {
     "board": "sensebox:samd:sb",
-    "port": "COM8",
-    "sketch": "examples\\SDCard\\SDCard.ino",
+    "port": "COM7",
+    "sketch": "examples\\blockly-test\\blockly-test.ino",
     "configuration": "power=on"
 }
